requirement.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
faiss-cpu==1.13.0
fastapi==0.121.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3

mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.5
openai==2.8.1
packaging==25.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
fbgemm-gpu-genai
torchao
llama-cpp-python